// src/constants/status.enum.ts
/**
 * Copyright (C) 2023 ECUALEAD
 *
 * All Rights Reserved
 * Author: Reinier Millo Sánchez <rmillo@ecualead.com>
 *
 * This file is part of the Developer Server Package
 * It can't be copied and/or distributed without the express
 * permission of the author.
 */

/**
 * Predefined base status
 */
export enum SERVER_STATUS {
  SOFT_DELETE = -1,
  UNKNOWN = 0,
  DISABLED = 1,
  ENABLED = 2
}